The following methods (functions) are exported and available by using SDC.

SDC.gauss_lobatto_nodes — Method

Get n Gauss-Lobatto quadrature nodes, on interval [0-1].

SDC.make_SDC — Method

This function returns a function: an SDC solver which needs only the initial conditions as the arguments. Moreover, the quadrature and spectral integration matrix are built-in to the function, not calculated each run.

SDC.make_spectral_int_matrix — Method

Get the spectral integration matrix

SDC.sdc — Method

Notes for improvement:

The main usage for this function will probably be something like: for i in 1:numberoftimesteps SDC(...)

So I don't want to waste resources recomputing S everytime, or setting up all the equations.

Perhaps I should have a create_SDC function? This would be provided all parameters except the initial guess, and create an SDC function which only needs the initial values as arguments.

From a UX perspective, I like that they will have an SDC function whose only arguments are the initial data. That would be very easy to use. But I don't like that they have to call a function to create a function. That seems like it might be confusing.

Actually, I think it would be good to also have Δt as an argument to the user-SDC function. I can imagine a user might want to use variable timestep size (but they probably don't need to vary the number of collocation nodes, implicit/explicit f, etc).